// src/interfaces/Product.ts
import { Props as ProductButtonsProps } from "../components/ProductButtons"
import { Props as ProductCardProps } from "../components/ProductCard"
import { Props as ProductImageProps } from "../components/ProductImage"
import { Props as ProductTitleProps } from "../components/ProductTitle"


export interface Product {
  id: string
  title: string
  img?: string
}

export interface ProductContextProps {
  counter: number
  maxCount?: number
  product: Product
  increaseBy: (value: number) => void
}

export interface ProductCardHOCProps {
  ({ children, product }: ProductCardProps): JSX.Element
  Image: React.FC<ProductImageProps>
  Title: React.FC<ProductTitleProps>
  Buttons: React.FC<ProductButtonsProps>
}


export interface onChangeArgs {
  product: Product,
  count: number
}


export interface ProductInCart extends Product {
  count: number
}

export interface ShoppingCartState {
  [key: string]: ProductInCart
}

export interface InitialValues {
  count?: number,
  maxCount?: number
}

export interface ProductCardHandlers {
  count: number
  isMaxCountReached: boolean
  maxCount?: number
  product: Product

  increaseBy: (value: number) => void
  reset: () => void
}